add request parameters and clean up code
