+++ .gitignore
@@ -104,6 +104,3 @@ ENV/
 
 # keys
 privatekey*
-
-# vscode
-.vscode
