+++ hydroengine_service/digitwin_functions.py
@@ -1,5 +1,6 @@
 # coding: utf-8
 import numpy as np
+import pandas as pd
 import scipy.interpolate
 
 import ee
@@ -214,3 +215,153 @@ def compute_distortion(geometry):
     scaleY = ee.Number(1).divide(centerPlusOneY.distance(center3857))
     result = ee.Dictionary({"scaleX": scaleX, "scaleY": scaleY})
     return result
+
+KNOWN_MODELS = ('HYCOM', 'GLOSSIS')
+
+def submit_ecopath_jobs(scale, crs, model, t_start, n_periods):
+    """
+    submit ecopath job to calculate a monthly image for northsea currents with the following
+    args:
+        scale: scale of the model pixels in m
+        model: Either "HYCOM" or "GLOSSIS"
+        t_start: string containing start-date
+        crs: crs of images to be exported
+        n_periods: number of months to export from t_start
+    returns:
+        list of tasks that were submitted to earthengine
+    """
+
+    assert model in KNOWN_MODELS, f"model not in {KNOWN_MODELS}"
+
+    period = 'month'
+    t_stop = pd.Timestamp(t_start) + pd.DateOffset(months=n_periods)
+    bucket = 'hydro-engine-public'
+
+    extent = [
+        3463000.0001221001148224, # xmin
+        3117163.4356725001707673, # ymin
+        4313000.0001221001148224, # xmax
+        4257163.4356725001707673  # ymax
+    ]
+    geometry =  ee.Geometry.Polygon(
+        coords=[[
+            [extent[0], extent[1]],
+            [extent[0], extent[3]],
+            [extent[2], extent[3]],
+            [extent[2], extent[1]],
+            [extent[0], extent[1]]
+        ]], 
+        proj=crs, 
+        geodesic=False, 
+        maxError=1
+    )
+
+    # Get ImageCollections
+    hycom_currents = ee.ImageCollection("HYCOM/sea_water_velocity")
+    glossis_currents = ee.ImageCollection("projects/dgds-gee/glossis/currents")
+    
+    # Determine all period ranges 
+    def create_period_step(num_period):
+        return ee.DateRange(
+            ee.Date(t_start).advance(num_period, period),  # take i-th period
+            ee.Date(t_start).advance(ee.Number(num_period).add(1), period)  # until i+1th period
+        )
+
+    period_ranges = ee.List.sequence(0, n_periods - 1).map(create_period_step)
+
+    # Prepare the datasets
+    # make datasets more consistent
+    # Use top layer for velocities in hycom
+    def get_top_layer(img):
+        result = img \
+            .select(['velocity_u_0', 'velocity_v_0']) \
+            .rename(['velocity_u', 'velocity_v']) \
+            .float() \
+            .multiply(0.001)
+        
+        result = result.copyProperties(img)
+        result = result.set('system:time_start', img.get('system:time_start'))
+        return ee.Image(result)
+
+    # Apply the get top layer function to all images
+    hycom_currents = hycom_currents.filterDate(t_start, t_stop)
+    hycom_currents = hycom_currents.map(get_top_layer)
+
+    def get_compatible_glossis_currents(img):
+        result = img.rename(['velocity_u', 'velocity_v'])
+        result = result.resample('bicubic')
+        return result.set('system:time_start', img.get('system:time_start'))
+
+    glossis_currents = glossis_currents.map(get_compatible_glossis_currents)
+
+    def get_map(model, period_range):
+        models = {
+            "HYCOM": hycom_currents,
+            "glossis": glossis_currents
+        }
+
+        selected = models[model]
+        period_range = ee.DateRange(period_range)
+
+        selected = selected.filterDate(
+            period_range.start(),
+            period_range.end()
+        )
+
+        image_count = selected.size()
+        selected = selected.mean()
+        selected = selected.set('model', model)
+        selected = selected.set('tStart', period_range.start())
+        selected = selected.set('tStop', period_range.end())
+        selected = selected.set('imageCount', image_count)
+        selected = selected.set('units', 'm/s')
+        return selected
+    
+    def export_map(image, export_path, geometry, scale, crs):
+        task =  ee.batch.Export.image.toCloudStorage(
+            image=image,
+            description=export_path,
+            bucket=bucket,
+            fileNamePrefix=export_path,
+            region=geometry,
+            scale=scale,
+            crs=crs 
+        )
+        task.start()
+        return task
+
+    # Now for some annoying stuff, we need to make a list of all properties
+    # so that we can call export with local variables
+
+    def prep_export(period_range):
+        _t_start = ee.DateRange(period_range).start()
+        export_path = ee.String("ecopath/").cat(
+            ee.String(model).cat(_t_start.format())
+        )
+        im = get_map(model, period_range)
+        im = im.set('path', export_path)
+        im = im.set('tStart',  _t_start)
+        im = im.set('system:time_start', _t_start.millis())
+        im = im.set('period', period_range)
+        return im
+
+    images = period_ranges.map(prep_export)
+
+    # filter out empty images
+    images = ee.ImageCollection(images)
+    images = images.filter(ee.Filter.gt('imageCount', 0))
+
+    # sort images on path
+    images = images.sort("path")
+    # paths will shading images "path" property as a list
+    date_range = pd.date_range(start=t_start, periods=n_periods, freq=pd.DateOffset(months=1))
+    paths=[f"ecopath_{model}{dr.strftime('%Y-%m-%dT%X').replace(':', '_')}" for dr in date_range]
+
+    def export_path(path):
+        # find the image that corresponds to the path
+        im = images.filter(ee.Filter.eq("path", path)).first()
+        return export_map(im, path, geometry, scale, crs)
+
+    task_list = [export_path(path) for path in paths]
+        
+    return task_list
