+++ hydroengine_service/main.py
@@ -9,8 +9,6 @@ import logging
 import os
 
 import ee
-import numpy as np
-import geojson
 import flask_cors
 from flask import Flask
 from flask import request, Response
@@ -25,10 +23,7 @@ from hydroengine_service import dgds_blueprints
 from hydroengine_service import river_functions
 from hydroengine_service import dgds_functions
 
-from hydroengine_service import liwo_functions
 from hydroengine_service import digitwin_blueprints
-
-from hydroengine_service import ecopath_blueprints
 
 logger = logging.getLogger(__name__)
 
@@ -1256,7 +1251,6 @@ app.register_blueprint(v1, url_prefix="/")
 app.register_blueprint(liwo_blueprints.v1, url_prefix="/")
 app.register_blueprint(dgds_blueprints.v1, url_prefix="/")
 app.register_blueprint(digitwin_blueprints.v1, url_prefix="/")
-app.register_blueprint(ecopath_blueprints.v1, url_prefix="/v1/ecopath")
 
 
 if __name__ == '__main__':
