+++ requirements.txt
@@ -9,4 +9,4 @@ oauth2client==4.1.*
 Click==6.*
 numpy==1.*
 six>=1.13.0,<2dev
-pandas>=1.0.5,<2.*
+pandas~=1.0.5
