+++ tests/test_digitwin_functions.py
@@ -9,7 +9,7 @@ import ee
 from . import auth
 
 from hydroengine_service import config
-from hydroengine_service.ecopath_functions import submit_ecopath_jobs
+from hydroengine_service.digitwin_functions import submit_ecopath_jobs
 
 
 logger = logging.getLogger(__name__)
@@ -22,5 +22,6 @@ ee.Initialize(EE_CREDENTIALS, opt_url="https://earthengine-highvolume.googleapis
 
 class TestClient(unittest.TestCase):
     def test_ecopath_exports(self):
-        tasks = submit_ecopath_jobs()
+        tasks = submit_ecopath_jobs(30000, "EPSG:3004", "HYCOM", "2020-07-01", 2)
         assert type(tasks[0]) == ee.batch.Task
+        assert tasks[0].id is not None  # make sure task is started
